refactor: Update transaction mapping and controller to use new package structure and use case naming
