+++ backend/transaction/src/main/java/br/com/transaction/adapters/input/TransactionControllerAdapter.java
@@ -1,7 +1,7 @@
 package br.com.transaction.adapters.input;
 
-import br.com.transaction.adapters.input.dto.CreateTransactionRequest;
-import br.com.transaction.domain.service.CreateTransactionService;
+import br.com.transaction.adapters.input.dto.transaction.create.CreateTransactionRequest;
+import br.com.transaction.domain.usecase.transaction.create.CreateTransactionUseCase;
 import br.com.transaction.ports.input.TransactionControllerPort;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.RestController;
@@ -9,15 +9,15 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 public class TransactionControllerAdapter implements TransactionControllerPort {
 
-    private final CreateTransactionService createTransactionService;
+    private final CreateTransactionUseCase createTransactionUseCase;
 
-    public TransactionControllerAdapter(CreateTransactionService createTransactionService) {
-        this.createTransactionService = createTransactionService;
+    public TransactionControllerAdapter(CreateTransactionUseCase createTransactionUseCase) {
+        this.createTransactionUseCase = createTransactionUseCase;
     }
 
     @Override
     public ResponseEntity<Void> createNewTransaction(CreateTransactionRequest createTransactionRequest) {
-        createTransactionService.saveTransaction(createTransactionRequest);
+        createTransactionUseCase.saveTransaction(createTransactionRequest);
         return ResponseEntity.status(201).build();
     }
 }
